Refactored BO and models code.
