+++ scripts/bo.py
@@ -1,0 +1,422 @@
+import sys
+import os
+import space
+import random
+import models
+import operator
+import numpy as np
+import csv
+import json
+import copy
+import datetime
+from jsonschema import Draft4Validator, validators, exceptions
+from utility_functions import *
+from collections import defaultdict
+from random_scalarizations import random_scalarizations
+from prior_optimization import prior_guided_optimization
+
+def main(config, black_box_function=None, output_file=""):
+    """
+    Run design-space exploration using bayesian optimization.
+    :param config: dictionary containing all the configuration parameters of this optimization.
+    :param output_file: a name for the file used to save the dse results.
+    """
+    start_time = (datetime.datetime.now())
+    run_directory = config["run_directory"]
+    hypermapper_mode = config["hypermapper_mode"]["mode"]
+
+    # Start logging
+    log_file = deal_with_relative_and_absolute_path(run_directory, config["log_file"])
+    sys.stdout.change_log_file(log_file)
+    if (hypermapper_mode == 'client-server'):
+        sys.stdout.switch_log_only_on_file(True)
+
+    # Log the json configuration for this optimization
+    sys.stdout.write_to_logfile(str(config) + "\n")
+
+    # Create parameter space object and unpack hyperparameters from json
+    param_space = space.Space(config)
+    application_name = config["application_name"]
+    optimization_metrics = config["optimization_objectives"]
+    optimization_iterations = config["optimization_iterations"]
+    evaluations_per_optimization_iteration = config["evaluations_per_optimization_iteration"]
+    number_of_cpus = config["number_of_cpus"]
+    print_importances = config["print_parameter_importance"]
+    epsilon_greedy_threshold = config["epsilon_greedy_threshold"]
+    acquisition_function = config["acquisition_function"]
+    weight_sampling = config["weight_sampling"]
+    scalarization_method = config["scalarization_method"]
+    scalarization_key = config["scalarization_key"]
+    doe_type = config["design_of_experiment"]["doe_type"]
+    number_of_doe_samples = config["design_of_experiment"]["number_of_samples"]
+    model_type = config["models"]["model"]
+    optimization_method = config["optimization_method"]
+    time_budget = config["time_budget"]
+    input_params = param_space.get_input_parameters()
+    number_of_objectives = len(optimization_metrics)
+    objective_limits = {}
+    data_array = {}
+    fast_addressing_of_data_array = {}
+    objective_bounds = None
+    exhaustive_search_data_array = None
+    normalize_objectives = False
+    debug = False
+
+    if "feasible_output" in config:
+        feasible_output = config["feasible_output"]
+        feasible_output_name = feasible_output["name"]
+        enable_feasible_predictor = feasible_output["enable_feasible_predictor"]
+        enable_feasible_predictor_grid_search_on_recall_and_precision = feasible_output["enable_feasible_predictor_grid_search_on_recall_and_precision"]
+        feasible_predictor_grid_search_validation_file = feasible_output["feasible_predictor_grid_search_validation_file"]
+        feasible_parameter = param_space.get_feasible_parameter()
+        number_of_trees = config["models"]["number_of_trees"]
+
+    if (weight_sampling == "bounding_box"):
+        objective_bounds = {}
+        user_bounds = config["bounding_box_limits"]
+        if (len(user_bounds) == 2):
+            if (user_bounds[0] > user_bounds[1]):
+                user_bounds[0], user_bounds[1] = user_bounds[1], user_bounds[0]
+            for objective in optimization_metrics:
+                objective_bounds[objective] = user_bounds
+                objective_limits[objective] = user_bounds
+        elif (len(user_bounds) == number_of_objectives*2):
+            idx = 0
+            for objective in optimization_metrics:
+                objective_bounds[objective] = user_bounds[idx:idx+2]
+                if (objective_bounds[objective][0] > objective_bounds[objective][1]):
+                    objective_bounds[objective][0], objective_bounds[objective][1] = objective_bounds[objective][1], objective_bounds[objective][0]
+                objective_limits[objective] = objective_bounds[objective]
+                idx += 2
+        else:
+            print("Wrong number of bounding boxes, expected 2 or", 2*number_of_objectives, "got", len(user_bounds))
+            raise SystemExit
+    else:
+        for objective in optimization_metrics:
+            objective_limits[objective] = [float("inf"), float("-inf")]
+
+    if output_file == "":
+        output_data_file = config["output_data_file"]
+        if output_data_file == "output_samples.csv":
+            output_data_file = application_name + "_" + output_data_file
+    else:
+        output_data_file = output_file
+
+    exhaustive_search_data_array = None
+    exhaustive_search_fast_addressing_of_data_array = None
+    if hypermapper_mode == 'exhaustive':
+        exhaustive_file = config["hypermapper_mode"]["exhaustive_search_file"]
+        exhaustive_search_data_array, exhaustive_search_fast_addressing_of_data_array = param_space.load_data_file(exhaustive_file, debug=False, number_of_cpus=number_of_cpus)
+
+    # Check if some parameters are correctly defined
+    if hypermapper_mode == "default":
+        if black_box_function == None:
+            print("Error: the black box function must be provided")
+            raise SystemExit
+        if not callable(black_box_function):
+            print("Error: the black box function parameter is not callable")
+            raise SystemExit
+
+    if (model_type == "gaussian_process") and (acquisition_function == "TS"):
+        print("Error: The TS acquisition function with Gaussian Process models is still under implementation")
+        print("Using EI acquisition function instead")
+        config["acquisition_function"] = "EI"
+
+    if number_of_cpus > 1:
+        print("Warning: HyperMapper supports only sequential execution for now. Running on a single cpu.")
+        number_of_cpus = 1
+
+    # If priors are present, use prior-guided optimization
+    user_priors = False
+    for input_param in config["input_parameters"]:
+        if config["input_parameters"][input_param]["prior"] != "uniform":
+            if number_of_objectives == 1:
+                user_priors = True
+            else:
+                print("Warning: prior optimization does not work with multiple objectives yet, priors will be uniform")
+                config["input_parameters"][input_param]["prior"] = "uniform"
+
+    if user_priors:
+        bo_method = prior_guided_optimization
+    else:
+        bo_method = random_scalarizations
+        normalize_objectives = True
+
+    ### Resume previous optimization, if any
+    beginning_of_time = param_space.current_milli_time()
+    absolute_configuration_index = 0
+    doe_t0 = datetime.datetime.now()
+    if config["resume_optimization"] == True:
+        resume_data_file = config["resume_optimization_data"]
+
+        if not resume_data_file.endswith('.csv'):
+            print("Error: resume data file must be a CSV")
+            raise SystemExit
+        if resume_data_file == "output_samples.csv":
+            resume_data_file = application_name + "_" + resume_data_file
+
+        data_array, fast_addressing_of_data_array = param_space.load_data_file(resume_data_file, debug=False, number_of_cpus=number_of_cpus)
+        absolute_configuration_index = len(data_array[list(data_array.keys())[0]]) # get the number of points evaluated in the previous run
+        beginning_of_time = beginning_of_time - data_array[param_space.get_timestamp_parameter()[0]][-1] # Set the timestamp back to match the previous run
+        print("Resumed optimization, number of samples = %d ......." % absolute_configuration_index)
+
+    ### DoE phase
+    if absolute_configuration_index < number_of_doe_samples:
+        configurations = []
+        default_configuration = param_space.get_default_or_random_configuration()
+        str_data = param_space.get_unique_hash_string_from_values(default_configuration)
+        if str_data not in fast_addressing_of_data_array:
+            fast_addressing_of_data_array[str_data] = absolute_configuration_index
+            configurations.append(default_configuration)
+            absolute_configuration_index += 1
+
+        doe_configurations = []
+        if absolute_configuration_index < number_of_doe_samples:
+            doe_configurations = param_space.get_doe_sample_configurations(
+                                                                        fast_addressing_of_data_array,
+                                                                        number_of_doe_samples-absolute_configuration_index,
+                                                                        doe_type)
+        configurations += doe_configurations
+        print("Design of experiment phase, number of new doe samples = %d ......." % len(configurations))
+
+        doe_data_array = param_space.run_configurations(
+                                                        hypermapper_mode,
+                                                        configurations,
+                                                        beginning_of_time,
+                                                        black_box_function,
+                                                        exhaustive_search_data_array,
+                                                        exhaustive_search_fast_addressing_of_data_array,
+                                                        run_directory)
+        data_array = concatenate_data_dictionaries(
+                                                data_array,
+                                                doe_data_array,
+                                                param_space.input_output_and_timestamp_parameter_names)
+        absolute_configuration_index = number_of_doe_samples
+        iteration_number = 1
+    else:
+        iteration_number = absolute_configuration_index - number_of_doe_samples + 1
+
+    # If we have feasibility constraints, we must ensure we have at least one feasible and one infeasible sample before starting optimization
+    # If this is not true, continue design of experiment until the condition is met
+    if enable_feasible_predictor:
+        while are_all_elements_equal(data_array[feasible_parameter[0]]) and optimization_iterations > 0:
+            print("Warning: all points are either valid or invalid, random sampling more configurations.")
+            print("Number of doe samples so far:", absolute_configuration_index)
+            configurations = param_space.get_doe_sample_configurations(fast_addressing_of_data_array, 1, "random sampling")
+            new_data_array = param_space.run_configurations(
+                                                            hypermapper_mode,
+                                                            configurations,
+                                                            beginning_of_time,
+                                                            black_box_function,
+                                                            exhaustive_search_data_array,
+                                                            exhaustive_search_fast_addressing_of_data_array,
+                                                            run_directory)
+            data_array = concatenate_data_dictionaries(
+                                                        new_data_array,
+                                                        data_array,
+                                                        param_space.input_output_and_timestamp_parameter_names)
+            absolute_configuration_index += 1
+            optimization_iterations -= 1
+
+    # Create output file with explored configurations from resumed run and DoE
+    with open(deal_with_relative_and_absolute_path(run_directory, output_data_file), 'w') as f:
+        w = csv.writer(f)
+        w.writerow(param_space.get_input_output_and_timestamp_parameters())
+        tmp_list = [param_space.convert_types_to_string(j, data_array) for j in param_space.get_input_output_and_timestamp_parameters()]
+        tmp_list = list(zip(*tmp_list))
+        for i in range(len(data_array[optimization_metrics[0]])):
+            w.writerow(tmp_list[i])
+
+    for objective in optimization_metrics:
+        lower_bound = min(objective_limits[objective][0], min(data_array[objective]))
+        upper_bound = max(objective_limits[objective][1], max(data_array[objective]))
+        objective_limits[objective] = [lower_bound, upper_bound]
+    print("\nEnd of doe/resume phase, the number of evaluated configurations is: %d\n" %absolute_configuration_index)
+    sys.stdout.write_to_logfile(("End of DoE - Time %10.4f sec\n" % ((datetime.datetime.now() - doe_t0).total_seconds())))
+    if doe_type == "grid_search" and optimization_iterations > 0:
+        print("Warning: DoE is grid search, setting number of optimization iterations to 0")
+        optimization_iterations = 0
+
+    ### Main optimization loop
+    if evaluations_per_optimization_iteration > 1:
+        print("Warning, number of evaluations per iteration > 1")
+        print("HyperMapper bayesian optimization currently does not support multiple runs per iteration, setting evaluations per iteration to 1")
+
+    bo_t0 = datetime.datetime.now()
+    run_time = (datetime.datetime.now() - start_time).total_seconds() / 60
+    # run_time / time_budget < 1 if budget > elapsed time or budget == -1
+    if time_budget > 0:
+        print('starting optimization phase, limited to run for ', time_budget, ' minutes')
+    elif time_budget == 0:
+        print('Time budget cannot be zero. To not limit runtime set time_budget = -1')
+        sys.exit()
+
+    while iteration_number <= optimization_iterations and run_time / time_budget < 1:
+        print("Starting optimization iteration", iteration_number)
+        iteration_t0 = datetime.datetime.now()
+        model_t0 = datetime.datetime.now()
+        regression_models,_,_ = models.generate_mono_output_regression_models(
+                                                                            data_array,
+                                                                            param_space,
+                                                                            input_params,
+                                                                            optimization_metrics,
+                                                                            1.00,
+                                                                            config,
+                                                                            model_type=model_type,
+                                                                            number_of_cpus=number_of_cpus,
+                                                                            print_importances=print_importances,
+                                                                            normalize_objectives=normalize_objectives,
+                                                                            objective_limits=objective_limits)
+
+        classification_model = None
+        if enable_feasible_predictor:
+            classification_model,_,_ = models.generate_classification_model(application_name,
+                                                                            param_space,
+                                                                            data_array,
+                                                                            input_params,
+                                                                            feasible_parameter,
+                                                                            1.00,
+                                                                            config,
+                                                                            debug,
+                                                                            number_of_cpus=number_of_cpus,
+                                                                            data_array_exhaustive=exhaustive_search_data_array,
+                                                                            enable_feasible_predictor_grid_search_on_recall_and_precision=enable_feasible_predictor_grid_search_on_recall_and_precision,
+                                                                            feasible_predictor_grid_search_validation_file=feasible_predictor_grid_search_validation_file,
+                                                                            print_importances=print_importances)
+        model_t1 = datetime.datetime.now()
+        if (weight_sampling == "bounding_box"):
+            objective_weights = sample_weight_bbox(optimization_metrics, objective_bounds, 1)[0]
+        elif (weight_sampling == "flat"):
+            objective_weights = sample_weight_flat(optimization_metrics, 1)[0]
+        else:
+            print("Error: unrecognized option:", weight_sampling)
+            raise SystemExit
+
+        data_array_scalarization, _ = compute_data_array_scalarization(
+                                                                    data_array,
+                                                                    objective_weights,
+                                                                    objective_limits,
+                                                                    scalarization_method)
+        data_array[scalarization_key] = data_array_scalarization.tolist()
+
+        epsilon = random.uniform(0,1)
+        local_search_t0 = datetime.datetime.now()
+        if epsilon > epsilon_greedy_threshold:
+            best_configuration = bo_method(
+                                        config,
+                                        data_array,
+                                        param_space,
+                                        fast_addressing_of_data_array,
+                                        regression_models,
+                                        iteration_number,
+                                        objective_weights,
+                                        objective_limits,
+                                        classification_model)
+
+        else:
+            sys.stdout.write_to_logfile(str(epsilon) + " < " + str(epsilon_greedy_threshold) + " random sampling a configuration to run\n")
+            best_configuration = param_space.random_sample_configurations_without_repetitions(fast_addressing_of_data_array, 1)[0]
+        local_search_t1 = datetime.datetime.now()
+
+        configurations = [best_configuration]
+        str_data = param_space.get_unique_hash_string_from_values(best_configuration)
+        fast_addressing_of_data_array[str_data] = absolute_configuration_index
+        absolute_configuration_index += 1
+
+        black_box_function_t0 = datetime.datetime.now()
+        new_data_array = param_space.run_configurations(
+                                                        hypermapper_mode,
+                                                        configurations,
+                                                        beginning_of_time,
+                                                        black_box_function,
+                                                        exhaustive_search_data_array,
+                                                        exhaustive_search_fast_addressing_of_data_array,
+                                                        run_directory)
+        black_box_function_t1 = datetime.datetime.now()
+
+        with open(deal_with_relative_and_absolute_path(run_directory, output_data_file), 'a') as f:
+            w = csv.writer(f)
+            tmp_list = [param_space.convert_types_to_string(j, new_data_array) for j in list(param_space.get_input_output_and_timestamp_parameters())]
+            tmp_list = list(zip(*tmp_list))
+            for i in range(len(new_data_array[optimization_metrics[0]])):
+                w.writerow(tmp_list[i])
+
+        data_array = concatenate_data_dictionaries(
+                                                new_data_array,
+                                                data_array,
+                                                param_space.input_output_and_timestamp_parameter_names)
+        for objective in optimization_metrics:
+            lower_bound = min(objective_limits[objective][0], min(data_array[objective]))
+            upper_bound = max(objective_limits[objective][1], max(data_array[objective]))
+            objective_limits[objective] = [lower_bound, upper_bound]
+
+        iteration_number += 1
+        run_time = (datetime.datetime.now() - start_time).total_seconds() / 60
+
+        sys.stdout.write_to_logfile(("Model fitting time %10.4f sec\n" % ((model_t1 - model_t0).total_seconds())))
+        sys.stdout.write_to_logfile(("Local search time %10.4f sec\n" % ((local_search_t1 - local_search_t0).total_seconds())))
+        sys.stdout.write_to_logfile(("Black box function time %10.4f sec\n" % ((black_box_function_t1 - black_box_function_t0).total_seconds())))
+        sys.stdout.write_to_logfile(("Total iteration time %10.4f sec\n" % ((datetime.datetime.now() - iteration_t0).total_seconds())))
+    sys.stdout.write_to_logfile(("End of BO phase - Time %10.4f sec\n" % ((datetime.datetime.now() - bo_t0).total_seconds())))
+
+    print("End of Bayesian Optimization")
+    sys.stdout.write_to_logfile(("Total script time %10.2f sec\n" % ((datetime.datetime.now() - start_time).total_seconds())))
+
+if __name__ == "__main__":
+    if len(sys.argv) == 2:
+        parameters_file = sys.argv[1]
+    else :
+        print("Error: only one argument needed, the parameters json file.")
+
+    if parameters_file == "--help" or len(sys.argv) != 2:
+        print("################################################")
+        print("### Example: ")
+        print("### cd hypermapper")
+        print("### python3 scripts/hypermapper.py example_scenarios/spatial/BlackScholes_scenario.json")
+        print("################################################")
+        raise SystemExit
+
+    try:
+        initial_directory = os.environ['PWD']
+        hypermapper_home = os.environ['HYPERMAPPER_HOME']
+        os.chdir(hypermapper_home)
+    except:
+        hypermapper_home = "."
+        initial_directory = "."
+
+    if not parameters_file.endswith('.json'):
+        _, file_extension = os.path.splitext(parameters_file)
+        print("Error: invalid file name. \nThe input file has to be a .json file not a %s" %file_extension)
+        raise SystemExit
+    with open(parameters_file, 'r') as f:
+        config = json.load(f)
+
+    json_schema_file = 'scripts/schema.json'
+    with open(json_schema_file, 'r') as f:
+        schema = json.load(f)
+
+    try:
+        DefaultValidatingDraft4Validator = extend_with_default(Draft4Validator)
+        DefaultValidatingDraft4Validator(schema).validate(config)
+    except exceptions.ValidationError as ve:
+        print("Failed to validate json:")
+        print(ve)
+        raise SystemExit
+
+    run_directory = config["run_directory"]
+    if run_directory == ".":
+        run_directory = initial_directory
+        config["run_directory"] = run_directory
+    log_file = config["log_file"]
+    if log_file == "hypermapper_logfile.log":
+        log_file = deal_with_relative_and_absolute_path(run_directory, log_file)
+    sys.stdout = Logger(log_file)
+
+    main(config)
+
+    try:
+        os.chdir(hypermapper_pwd)
+    except:
+        pass
+
+    print("### End of bayesian optimization script.")
