+++ scripts/models.py
@@ -10,6 +10,197 @@ from utility_functions import *
 from scipy import stats
 import GPy
 
+class RFModel(RandomForestRegressor):
+    """
+    Implementation of our adapted RF model. We extend scikit-learn's RF implementation to
+    implement the adapted RF model proposed by Hutter et al.: https://arxiv.org/abs/1211.0906
+    """
+    def __init__(self, **kwargs):
+        RandomForestRegressor.__init__(self, **kwargs)
+        self.tree_means_per_leaf = []
+        self.tree_vars_per_leaf = []
+
+    def set_means_per_leaf(self, samples, leaf_per_sample):
+        """
+        Compute the mean value for each leaf in the forest.
+        :param samples: list with the value of each sample used to build the forest.
+        :param leaf_per_sample: matrix with dimensions number_of_trees * number_of_samples. Stores the leaf each sample fell into for each tree.
+        :return: list of number_of_trees dictionaries. Each dictionary contains the means for each leaf in a tree.
+        """
+        number_of_trees, number_of_samples = leaf_per_sample.shape
+        for tree_idx in range(number_of_trees):
+            leaf_means = defaultdict(int)
+            leaf_sample_count = defaultdict(int)
+            for sample_idx in range(number_of_samples):
+                leaf = leaf_per_sample[tree_idx, sample_idx]
+                leaf_sample_count[leaf] += 1
+                leaf_means[leaf] += samples[sample_idx]
+            for leaf in leaf_sample_count.keys():
+                leaf_means[leaf] = leaf_means[leaf]/leaf_sample_count[leaf]
+            self.tree_means_per_leaf.append(leaf_means)
+
+    def get_means_per_leaf(self):
+        return self.tree_means_per_leaf
+
+    def set_vars_per_leaf(self, samples, leaf_per_sample):
+        """
+        Compute the variance for each leaf in the forest.
+        :param samples: list with the value of each sample used to build the forest.
+        :param leaf_per_sample: matrix with dimensions number_of_trees * number_of_samples. Stores the leaf each sample fell into for each tree.
+        :return: list of number_of_trees dictionaries. Each dictionary contains the variance for each leaf in a tree.
+        """
+        number_of_trees, number_of_samples = leaf_per_sample.shape
+        for tree_idx in range(number_of_trees):
+            samples_per_leaf = defaultdict(list)
+            for sample_idx in range(number_of_samples):
+                leaf = leaf_per_sample[tree_idx, sample_idx]
+                samples_per_leaf[leaf].append(samples[sample_idx])
+
+            leaf_vars = {}
+            for leaf in samples_per_leaf.keys():
+                if len(samples_per_leaf[leaf]) > 1:
+                    leaf_vars[leaf] = np.var(samples_per_leaf[leaf], ddof=1)
+                else:
+                    leaf_vars[leaf] = 0
+                # leaf_vars[leaf] = max(leaf_vars[leaf], 0.01) # This makes HyperMapper exploit too much. We will revisit this.
+            self.tree_vars_per_leaf.append(leaf_vars)
+
+    def get_vars_per_leaf(self):
+        return self.tree_vars_per_leaf
+
+    def get_samples_per_node(self, tree, leaf_per_sample):
+        """
+        Compute which samples passed through each node in a tree.
+        :param tree: sklearn regression tree.
+        :param leaf_per_sample: matrix with dimensions number_of_trees * number_of_samples. Stores the leaf each sample fell into for each tree.
+        :return: list of lists. Each internal list contains which samples went through the node represented by the index in the outer list.
+        """
+        node_count = tree.tree_.node_count
+        samples_per_node = [[] for i in range(node_count)]
+        for sample_idx in range(len(leaf_per_sample)):
+            leaf = leaf_per_sample[sample_idx]
+            samples_per_node[leaf].append(sample_idx)
+
+        parents = [-1]*node_count
+        left_children = tree.tree_.children_left
+        right_children = tree.tree_.children_right
+        for node_idx in range(node_count):
+            if left_children[node_idx] != -1:
+                parents[left_children[node_idx]] = node_idx
+            if right_children[node_idx] != -1:
+                parents[right_children[node_idx]] = node_idx
+
+        for node_idx in range(node_count-1, -1, -1):
+            parent = parents[node_idx]
+            if parent != -1:
+                samples_per_node[parent] += samples_per_node[node_idx]
+        return samples_per_node
+
+    def get_node_bounds(self, samples, data_array, threshold):
+        """
+        Compute the lower and upper bounds used to make a splitting decision at a tree node.
+        :param samples: list containing the indices of all samples that went through the node.
+        :param data_array: list containing the values of one parameter for all of the samples from the data.
+        :param threshold: original threshold used to split the node.
+        :return: lower and upper bound that were used to compute the split.
+        """
+        lower_bound = float('-inf')
+        upper_bound = float('inf')
+        for sample in samples:
+            sample_value = data_array[sample]
+            if sample_value <= threshold:
+                lower_bound = max(lower_bound, data_array[sample])
+            else:
+                upper_bound = min(upper_bound, data_array[sample])
+
+        return lower_bound, upper_bound
+
+    def get_leaves_per_sample(self, bufferx, param_space):
+        """
+        Compute in which leaf each sample falls into for each tree.
+        :param bufferx: list containing the samples.
+        :param param_space: parameter space object for the current application.
+        :return: array containing the corresponding leaf of each tree for each sample.
+        """
+        leaf_per_sample = np.array([tree.apply(bufferx) for tree in self])
+        return leaf_per_sample
+
+    def fit_RF(self, X, y, data_array=None, param_space=None, **kwargs):
+        """
+        Fit the adapted RF model. If the data_array and param_space parameters are not provided
+        a standard scikit-learn RF model will be fitted instead.
+        :param X: the training data for the RF model.
+        :param y: the training data labels for the RF model.
+        :param data_array: a dictionary containing previously explored points and their function values.
+        :param param_space: parameter space object for the current application.
+        """
+        self.fit(X, y, **kwargs)
+
+        # If data_array and param_space are provided, fit the adapted RF
+        if (data_array is not None) and (param_space is not None):
+            bufferx = data_dictionary_to_tuple(data_array, list(data_array.keys()))
+            leaf_per_sample = self.get_leaves_per_sample(bufferx, param_space)
+            self.set_means_per_leaf(y, leaf_per_sample)
+            self.set_vars_per_leaf(y, leaf_per_sample)
+            new_features = list(data_array.keys())
+            for tree_idx, tree in enumerate(self):
+                samples_per_node = self.get_samples_per_node(tree, leaf_per_sample[tree_idx, :])
+
+                left_children = tree.tree_.children_left
+                right_children = tree.tree_.children_right
+                for node_idx in range(tree.tree_.node_count):
+                    if left_children[node_idx] == right_children[node_idx]: # If both children are equal, this is a leaf in the tree
+                        continue
+                    feature = new_features[tree.tree_.feature[node_idx]]
+                    threshold = tree.tree_.threshold[node_idx]
+
+                    lower_bound, upper_bound = self.get_node_bounds(samples_per_node[node_idx], data_array[feature], threshold)
+                    new_split = stats.uniform.rvs(loc=lower_bound, scale=upper_bound - lower_bound)
+                    tree.tree_.threshold[node_idx] = new_split
+
+
+    def compute_rf_prediction(self, leaf_per_sample):
+        """
+        Compute the forest prediction for a list of samples based on the mean of the leaves in each tree.
+        :param leaf_per_sample: matrix with dimensions number_of_trees * number_of_samples. Stores the leaf each sample fell into for each tree.
+        :return: list containing the mean of each sample.
+        """
+        number_of_trees, number_of_samples = leaf_per_sample.shape
+        sample_means = np.zeros(number_of_samples)
+        tree_means_per_leaf = self.get_means_per_leaf()
+        for tree_idx in range(number_of_trees):
+            for sample_idx in range(number_of_samples):
+                sample_leaf = leaf_per_sample[tree_idx, sample_idx]
+                sample_means[sample_idx] += tree_means_per_leaf[tree_idx][sample_leaf]/number_of_trees
+        return sample_means
+
+    def compute_rf_prediction_variance(self, leaf_per_sample, sample_means):
+        """
+        Compute the forest prediction variance for a list of samples based on the mean and variances of the leaves in each tree.
+        The variance is computed as proposed by Hutter et al. in https://arxiv.org/pdf/1211.0906.pdf.
+        :param leaf_per_sample: matrix with dimensions number_of_trees * number_of_samples. Stores the leaf each sample fell into for each tree.
+        :param sample_means: list containing the mean of each sample.
+        :return: list containing the variance of each sample.
+        """
+        number_of_trees, number_of_samples = leaf_per_sample.shape
+        mean_of_the_vars = np.zeros(number_of_samples)
+        var_of_the_means = np.zeros(number_of_samples)
+        sample_vars = np.zeros(number_of_samples)
+        tree_means_per_leaf = self.get_means_per_leaf()
+        tree_vars_per_leaf = self.get_vars_per_leaf()
+        for sample_idx in range(number_of_samples):
+            for tree_idx in range(number_of_trees):
+                sample_leaf = leaf_per_sample[tree_idx, sample_idx]
+                mean_of_the_vars[sample_idx] += tree_vars_per_leaf[tree_idx][sample_leaf]/number_of_trees
+                var_of_the_means[sample_idx] += (tree_means_per_leaf[tree_idx][sample_leaf]**2)/number_of_trees
+
+            var_of_the_means[sample_idx] = abs(var_of_the_means[sample_idx] - sample_means[sample_idx]**2)
+            sample_vars[sample_idx] = mean_of_the_vars[sample_idx] + var_of_the_means[sample_idx]
+            if sample_vars[sample_idx] == 0:
+                sample_vars[sample_idx] = 0.00001
+
+        return sample_vars
+
 
 def generate_multi_output_regression_model(data_array
                                , param_space
@@ -55,7 +246,7 @@ def generate_multi_output_regression_model(data_array
     if len (Y_test) == 0:
         Y_test = Y[:]
 
-    regressor = customRegressor(n_estimators=n_estimators, max_features = max_features, n_jobs=1, bootstrap=False, min_samples_split=5) 
+    regressor = customRegressor(n_estimators=n_estimators, max_features = max_features, n_jobs=1, bootstrap=False, min_samples_split=5)
     regressor.fit(X_train, y_train)
 
     if print_importances:
@@ -70,10 +261,13 @@ def generate_mono_output_regression_models(data_array
                                , Xcols
                                , Ycols
                                , learn_ratio
+                               , config
                                , debug=False
                                , model_type="random_forest"
                                , number_of_cpus=0
                                , print_importances=False
+                               , normalize_objectives=False
+                               , objective_limits=None
                                , **model_kwargs):
     """
     Fit a regression model, supported model types are Random Forest and Gaussian Process.
@@ -83,13 +277,29 @@ def generate_mono_output_regression_models(data_array
     :param Ycols: the names of the output labels used for training.
     :param learn_ratio: percentage of the input vectors used for training. A part of it maybe left over for cross validation.
     :param debug: is debugging mode enabled?
-    :param n_estimators: number of trees.
-    :param max_features: this is a parameter of the Random Forest. It decides how many feature to randomize.
-    :param model_type:
-    :param number_of_cpus:
+    :param model_type: type of model to create. Either random_forest or gaussian_process.
+    :param number_of_cpus: number of cpus to use in the models.
+    :param print_importances: whether to print the importance of each parameter according to the model.
+    :param normalize_objectives: whether to normalize the objective data before fitting the model.
+    :param objective_limits: limits for normalizing the objective data.
     :return: 3 variables: the classifier, X_test , Y_test.
     """
     start_time = datetime.datetime.now()
+
+    normalized_data_array = copy.deepcopy(data_array)
+    if normalize_objectives:
+        if objective_limits is not None:
+            for col in Ycols:
+                # Both limits are the same only if all elements in the array are equal. This causes the normalization to divide by 0.
+                # We cannot optimize an objective when all values are the same, so we set it to 0
+                if objective_limits[col][1] == objective_limits[col][0]:
+                    normalized_col = [0]*len(normalized_data_array[col])
+                else:
+                    normalized_col = [(x - objective_limits[col][0]) \
+                                            /(objective_limits[col][1] - objective_limits[col][0]) for x in normalized_data_array[col]]
+                normalized_data_array[col] = normalized_col
+        else:
+            sys.stdout.write_to_logfile("Warning: no limits provided, skipping objective normalization.\n")
 
     if param_space.get_input_normalization_flag() is True:
         compute_mean_and_std(data_array, param_space)
@@ -106,7 +316,7 @@ def generate_mono_output_regression_models(data_array
         X_test = X[:]
 
     for i, Ycol in enumerate(Ycols):
-        Y = data_array[Ycol]
+        Y = normalized_data_array[Ycol]
         y_train      = Y[0:learn_size]
         Y_test[Ycol] = Y[learn_size:]
         if len (Y_test[Ycol]) == 0:
@@ -132,8 +342,12 @@ def generate_mono_output_regression_models(data_array
             if print_importances:
                 print("Feature importance is currently not supported with GPs")
         elif model_type == "random_forest":
-            regressor[Ycol] = RandomForestRegressor(**model_kwargs) 
-            regressor[Ycol].fit(X_train, y_train)
+            n_estimators = config["models"]["number_of_trees"]
+            max_features = config["models"]["max_features"]
+            bootstrap = config["models"]["bootstrap"]
+            min_samples_split = config["models"]["min_samples_split"]
+            regressor[Ycol] = RFModel(n_estimators=n_estimators, max_features=max_features, bootstrap=bootstrap, min_samples_split=min_samples_split, **model_kwargs)
+            regressor[Ycol].fit_RF(X_train, y_train, data_array=preprocessed_data_array, param_space=param_space)
             if print_importances:
                 parameter_importances = compute_parameter_importance(regressor[Ycol], Xcols, param_space)
                 print("Regression model on " + str(Ycol) + ". Features names: " + str(Xcols) + ", feature importances: " + str(parameter_importances))
@@ -150,6 +364,7 @@ def generate_classification_model(application_name
                                   , Xcols
                                   , Ycols
                                   , learn_ratio
+                                  , config
                                   , debug=False
                                   , n_estimators=15
                                   , max_features=0.5
@@ -162,7 +377,7 @@ def generate_classification_model(application_name
     """
     Fit a Random Forest model (for now it is Random Forest but in the future we will host more models here (e.g. GPs and lattices).
     :param application_name: the name of the application given by the json file.
-    :param param_space:
+    :param param_space: parameter space object for the current application.
     :param data_array: the data to use for training.
     :param Xcols: the names of the input features used for training.
     :param Ycols: the names of the output labels used for training.
@@ -247,7 +462,7 @@ def compute_recall_and_precision_on_RF_hyperparameters(dataset, param_space, X_t
     Notice that using the accuracy instead of the recall would be the wrong thing to do because we care more about the
     fact that we want to classify precisely the true class (we care less of the false class).
     :param dataset: the test data to use to check the recall and precision.
-    :param param_space:
+    :param param_space: parameter space object for the current application.
     :param X_train: the features data used to train the Random Forests.
     :param y_train: the labels data used to train the Random Forests.
     :return:
@@ -303,6 +518,14 @@ def compute_recall_and_precision_on_RF_hyperparameters(dataset, param_space, X_t
     print("#######################################################################")
 
 def compute_parameter_importance(model, input_params, param_space):
+    """
+    Compute the importance for each input parameter in a RF model.
+    For now, importance can only be computed for RF models.
+    :param model: a RF model.
+    :param input_params: the input parameters whose importance we want to compute
+    :param param_space: parameter space object for the current application.
+    :return: list containing the importance of each input parameter.
+    """
     parameter_importances = [0]*len(input_params)
     categorical_parameters = param_space.get_input_categorical_parameters_objects(input_params)
     feature_idx = 0
@@ -315,6 +538,61 @@ def compute_parameter_importance(model, input_params, param_space):
             parameter_importances[idx] = model.feature_importances_[feature_idx]
             feature_idx += 1
     return parameter_importances
+
+def compute_model_mean_and_uncertainty(bufferx, model, model_type, param_space, var=False):
+    """
+    Compute the predicted mean and uncertainty (either standard deviation or variance) for a number of points.
+    :param bufferx: list containing points to predict.
+    :param model: model to use for the prediction.
+    :param model_type: the type of the model being used. For now, either random_forest or gaussian_process.
+    :param param_space: parameter space object for the current application.
+    :param var: whether to compute variance or standard deviation.
+    :return: the predicted mean and uncertaitny for each point.
+    """
+    if model_type == "random_forest":
+        mean, uncertainty = compute_rf_prediction_mean_and_uncertainty(bufferx, model, param_space, var=var)
+    if model_type == "gaussian_process":
+        mean, uncertainty = compute_gp_prediction_mean_and_uncertainty(bufferx, model, param_space, var=var)
+    return mean, uncertainty
+
+def compute_rf_prediction_mean_and_uncertainty(bufferx, model, param_space, var=False):
+    """
+    Compute the predicted mean and uncertainty (either standard deviation or variance) for a number of points with a RF model.
+    :param bufferx: list containing points to predict.
+    :param model: model to use for the prediction.
+    :param param_space: parameter space object for the current application.
+    :param var: whether to compute variance or standard deviation.
+    :return: the predicted mean and uncertaitny for each point.
+    """
+    prediction_means = {}
+    prediction_uncertainty = {}
+    normalized_bufferx = preprocess_data_buffer(bufferx, param_space)
+    for objective in model:
+        leaf_per_sample = model[objective].get_leaves_per_sample(normalized_bufferx, param_space)
+        prediction_means[objective] = model[objective].compute_rf_prediction(leaf_per_sample)
+        prediction_variances = model[objective].compute_rf_prediction_variance(leaf_per_sample, prediction_means[objective])
+        if var:
+            prediction_uncertainty[objective] = prediction_variances
+        else:
+            prediction_uncertainty[objective] = np.sqrt(prediction_variances)
+
+    return prediction_means, prediction_uncertainty
+
+def sample_model_posterior(bufferx, model, model_type, param_space):
+    """
+    Compute the value of a number of points according to a function sapmled from the model posterior.
+    :param bufferx: list containing points to predict.
+    :param model: model to use for the prediction.
+    :param model_type: the type of the model being used. For now, either random_forest or gaussian_process.
+    :param param_space: parameter space object for the current application.
+    :return: the predicted value of each point.
+    """
+    if model_type == "random_forest":
+        model_predictions = model_prediction(bufferx, model, param_space)
+    elif model_type == "gaussian_process":
+        model_predictions = sample_gp_posterior(bufferx, model, param_space)
+    return model_predictions
+
 
 def parallel_model_prediction(model, bufferx, param_space, debug=False, number_of_cpus=0):
     """
@@ -332,7 +610,7 @@ def parallel_model_prediction(model, bufferx, param_space, debug=False, number_o
         return domain_decomposition_and_parallel_computation(debug, mono_output_model_prediction, concatenate_function_prediction, bufferx, number_of_cpus, model, param_space)
     else:
         return domain_decomposition_and_parallel_computation(debug, multi_output_model_prediction, concatenate_function_prediction, bufferx, number_of_cpus, model, param_space)
-    
+
 
 def multi_output_model_prediction(bufferx, model, param_space):
     """
@@ -388,22 +666,6 @@ def model_probabilities(bufferx, model, param_space):
         Cresult[parameter] = model[parameter].predict_proba(normalized_bufferx)
     return Cresult
 
-def tree_predictions(bufferx, model, param_space):
-    normalized_bufferx = preprocess_data_buffer(bufferx, param_space)
-    optimization_metrics = param_space.get_optimization_parameters()
-    tree_predictions = {}
-    for metric in optimization_metrics:
-        tree_predictions[metric] = np.array([tree.predict(normalized_bufferx) for tree in model[metric]])
-    return tree_predictions
-
-def get_leaves_per_sample(bufferx, model, param_space):
-    normalized_bufferx = preprocess_data_buffer(bufferx, param_space)
-    optimization_metrics = param_space.get_optimization_parameters()
-    leaf_per_sample = {}
-    for metric in optimization_metrics:
-        leaf_per_sample[metric] = np.array([tree.apply(normalized_bufferx) for tree in model[metric]])
-    return leaf_per_sample
-
 def concatenate_function_prediction(results_parallel) :
     """
     Concatenate the results of parallel predictions into a single data dictionary.
@@ -450,7 +712,7 @@ def preprocess_data_array(data_array, param_space, input_params):
             param_size = parameter_objects[param].get_size()
             preprocessed_data_array[param] = []
             for param_value in data_array[param]:
-                preprocessed_data_array[param].append(param_value_list.index(param_value)/param_size) 
+                preprocessed_data_array[param].append(param_value_list.index(param_value)/param_size)
         elif param_space.get_input_normalization_flag() is True:
             X = np.array(data_array[param], dtype=np.float64)
             mean = param_space.get_parameter_mean(param)
@@ -480,218 +742,20 @@ def compute_mean_and_std(data_array, param_space):
         param_space.set_parameter_mean(param, mean)
         param_space.set_parameter_std(param, std)
 
-def get_mean_per_leaf(samples, leaf_per_sample):
+def compute_gp_prediction_mean_and_uncertainty(bufferx, model, param_space, var=False):
     """
-    Compute the mean value for each leaf in the forest.
-    :param samples: list with the value of each sample used to build the forest.
-    :param leaf_per_sample: matrix with dimensions number_of_trees * number_of_samples. Stores the leaf each sample fell into for each tree.
-    :return: list of number_of_trees dictionaries. Each dictionary contains the means for each leaf in a tree.
-    """
-    number_of_trees, number_of_samples = leaf_per_sample.shape
-    tree_means_per_leaf = []
-    for tree_idx in range(number_of_trees):
-        leaf_means = defaultdict(int)
-        leaf_sample_count = defaultdict(int)
-        for sample_idx in range(number_of_samples):
-            leaf = leaf_per_sample[tree_idx, sample_idx]
-            leaf_sample_count[leaf] += 1
-            leaf_means[leaf] += samples[sample_idx]
-        for leaf in leaf_sample_count.keys():
-            leaf_means[leaf] = leaf_means[leaf]/leaf_sample_count[leaf]
-        tree_means_per_leaf.append(leaf_means)
-    return tree_means_per_leaf
-
-def get_var_per_leaf(samples, leaf_per_sample):
-    """
-    Compute the variance for each leaf in the forest.
-    :param samples: list with the value of each sample used to build the forest.
-    :param leaf_per_sample: matrix with dimensions number_of_trees * number_of_samples. Stores the leaf each sample fell into for each tree.
-    :return: list of number_of_trees dictionaries. Each dictionary contains the variance for each leaf in a tree.
-    """
-    number_of_trees, number_of_samples = leaf_per_sample.shape
-    tree_vars_per_leaf = []
-    for tree_idx in range(number_of_trees):
-        samples_per_leaf = defaultdict(list)
-        for sample_idx in range(number_of_samples):
-            leaf = leaf_per_sample[tree_idx, sample_idx]
-            samples_per_leaf[leaf].append(samples[sample_idx])
-
-        leaf_vars = {}
-        for leaf in samples_per_leaf.keys():
-            if len(samples_per_leaf[leaf]) > 1:
-                leaf_vars[leaf] = np.var(samples_per_leaf[leaf], ddof=1)
-            else:
-                leaf_vars[leaf] = 0
-            # leaf_vars[leaf] = max(leaf_vars[leaf], 0.01) # This makes HyperMapper exploit too much. We will revisit this.
-        tree_vars_per_leaf.append(leaf_vars)
-
-    return tree_vars_per_leaf
-
-def compute_rf_prediction(leaf_per_sample, leaf_means):
-    """
-    Compute the forest prediction for a list of samples based on the mean of the leaves in each tree.
-    :param leaf_per_sample: matrix with dimensions number_of_trees * number_of_samples. Stores the leaf each sample fell into for each tree.
-    :param leaf_means: list of number_of_trees dictionaries. Each dictionary contains the mean for each leaf in a tree.
-    :return: list containing the mean of each sample.
-    """
-    number_of_trees, number_of_samples = leaf_per_sample.shape
-    sample_means = np.zeros(number_of_samples)
-    for tree_idx in range(number_of_trees):
-        for sample_idx in range(number_of_samples):
-            sample_leaf = leaf_per_sample[tree_idx, sample_idx]
-            sample_means[sample_idx] += leaf_means[tree_idx][sample_leaf]/number_of_trees
-    return sample_means
-
-def compute_rf_prediction_variance(leaf_per_sample, sample_means, leaf_means, leaf_vars):
-    """
-    Compute the forest prediction variance for a list of samples based on the mean and variances of the leaves in each tree.
-    The variance is computed as proposed by Hutter et al. in https://arxiv.org/pdf/1211.0906.pdf.
-    :param leaf_per_sample: matrix with dimensions number_of_trees * number_of_samples. Stores the leaf each sample fell into for each tree.
-    :param sample_means: list containing the mean of each sample.
-    :param leaf_means: list of number_of_trees dictionaries. Each dictionary contains the mean for each leaf in a tree.
-    :param leaf_vars: list of number_of_trees dictionaries. Each dictionary contains the variance for each leaf in a tree.
-    :return: list containing the variance of each sample.
-    """
-    number_of_trees, number_of_samples = leaf_per_sample.shape
-    mean_of_the_vars = np.zeros(number_of_samples)
-    var_of_the_means = np.zeros(number_of_samples)
-    sample_vars = np.zeros(number_of_samples)
-    for sample_idx in range(number_of_samples):
-        for tree_idx in range(number_of_trees):
-            sample_leaf = leaf_per_sample[tree_idx, sample_idx]
-            mean_of_the_vars[sample_idx] += leaf_vars[tree_idx][sample_leaf]/number_of_trees
-            var_of_the_means[sample_idx] += (leaf_means[tree_idx][sample_leaf]**2)/number_of_trees
-
-        var_of_the_means[sample_idx] = abs(var_of_the_means[sample_idx] - sample_means[sample_idx]**2)
-        sample_vars[sample_idx] = mean_of_the_vars[sample_idx] + var_of_the_means[sample_idx]
-        if sample_vars[sample_idx] == 0:
-            sample_vars[sample_idx] = 0.00001
-
-    return sample_vars
-
-
-def get_samples_per_node(tree, leaf_per_sample):
-    """
-    Compute which samples passed through each node in a tree.
-    :param tree: sklearn regression tree.
-    :param leaf_per_sample: matrix with dimensions number_of_trees * number_of_samples. Stores the leaf each sample fell into for each tree.
-    :return: list of lists. Each internal list contains which samples went through the node represented by the index in the outer list.
-    """
-    node_count = tree.tree_.node_count
-    samples_per_node = [[] for i in range(node_count)]
-    for sample_idx in range(len(leaf_per_sample)):
-        leaf = leaf_per_sample[sample_idx]
-        samples_per_node[leaf].append(sample_idx)
-
-    parents = [-1]*node_count
-    left_children = tree.tree_.children_left
-    right_children = tree.tree_.children_right
-    for node_idx in range(node_count):
-        if left_children[node_idx] != -1:
-            parents[left_children[node_idx]] = node_idx
-        if right_children[node_idx] != -1:
-            parents[right_children[node_idx]] = node_idx
-
-    for node_idx in range(node_count-1, -1, -1):
-        parent = parents[node_idx]
-        if parent != -1:
-            samples_per_node[parent] += samples_per_node[node_idx]
-    return samples_per_node
-
-def get_node_bounds(samples, data_array, threshold):
-    """
-    Compute the lower and upper bounds used to make a splitting decision at a tree node.
-    :param samples: list containing the indices of all samples that went through the node.
-    :param data_array: list containing the values of one parameter for all of the samples from the data.
-    :param threshold: original threshold used to split the node.
-    :return: lower and upper bound that were used to compute the split.
-    """
-    lower_bound = float('-inf')
-    upper_bound = float('inf')
-    for sample in samples:
-        sample_value = data_array[sample]
-        if sample_value <= threshold:
-            lower_bound = max(lower_bound, data_array[sample])
-        else:
-            upper_bound = min(upper_bound, data_array[sample])
-
-    return lower_bound, upper_bound
-
-def transform_rf_using_uniform_splits(regression_models, data_array, param_space):
-    """
-    Change the splitting thresholds from a random forest regressor. Thresholds are changed from (upper_bound + lower_bound)/2
-    to a uniformly sampled split in the range (lower_bound, upper_bound).
-    This splitting approach was proposed by Hutter et al.: https://arxiv.org/pdf/1211.0906.pdf
-    :param regression_models: sklearn random forest regressor. The model to adapt.
-    :param data_array: dictionary containing the points evaluated so far.
-    :param input_params: the input parameters of the regressor.
-    :return: regressor with changed splitting rules.
-    """
-    input_params = param_space.get_input_parameters()
-    bufferx = data_dictionary_to_tuple(data_array, input_params)
-    leaf_per_sample = get_leaves_per_sample(bufferx, regression_models, param_space)
-    preprocessed_data_array = preprocess_data_array(data_array, param_space, input_params)
-    new_features = list(preprocessed_data_array.keys())
-    for objective in regression_models:
-        for tree_idx, tree in enumerate(regression_models[objective]):
-            samples_per_node = get_samples_per_node(tree, leaf_per_sample[objective][tree_idx, :])
-
-            left_children = tree.tree_.children_left
-            right_children = tree.tree_.children_right
-            for node_idx in range(tree.tree_.node_count):
-                if left_children[node_idx] == right_children[node_idx]: # If both children are equal, this is a leaf in the tree
-                    continue
-                feature = new_features[tree.tree_.feature[node_idx]]
-                threshold = tree.tree_.threshold[node_idx]
-
-                lower_bound, upper_bound = get_node_bounds(samples_per_node[node_idx], preprocessed_data_array[feature], threshold)
-                new_split = stats.uniform.rvs(loc=lower_bound, scale=upper_bound - lower_bound)
-
-                tree.tree_.threshold[node_idx] = new_split
-    return regression_models
-
-def compute_probability_from_model(
-                                model_means,
-                                model_stds,
-                                param_space,
-                                objective_weights,
-                                threshold,
-                                compute_bad=True):
-    """
-    Compute the probability of a configuration being good or bad according to the model.
-    :param model_means: predicted means of the model for each configuration.
-    :param model_means: predicted std of the model for each configuration.
-    :param param_space: Space object for the optimization problem.
-    :param objective_weights: objective weights for multi-objective optimization. Not implemented yet.
-    :param threshold: threshold on objective values separating good points and bad points.
-    :param compute_bad: whether to compute the probability of being good or bad.
-    """
-    optimization_parameters = param_space.get_optimization_parameters()
-    probabilities = np.ones(len(model_means[optimization_parameters[0]]))
-
-    for parameter in optimization_parameters:
-        parameter_means = model_means[parameter]
-        parameter_stds = model_stds[parameter]
-        if compute_bad:
-            p = 1 - stats.norm.cdf((threshold[parameter] - parameter_means)/parameter_stds)
-        else:
-            p = stats.norm.cdf((threshold[parameter] - parameter_means)/parameter_stds)
-        probabilities *= p**objective_weights[parameter]
-
-    return probabilities
-
-def compute_gp_prediction_mean_and_std(bufferx, model, param_space):
-    """
-    Compute the mean and std prediction of a GP model for a list of points.
+    Compute the predicted mean and uncertainty (either standard deviation or variance) for a number of points with a GP model.
     :param bufferx: list containing points to predict.
-    :param model: the GP regression model.
+    :param model: model to use for the prediction.
     :param param_space: parameter space object for the current application.
-    :return: mean and std predictions 
+    :param var: whether to compute variance or standard deviation.
+    :return: the predicted mean and uncertaitny for each point
     """
     normalized_bufferx = preprocess_data_buffer(bufferx, param_space)
     normalized_bufferx = np.array(normalized_bufferx)
     means = {}
     stds = {}
+    uncertainty = {}
     for parameter in model:
         means[parameter], stds[parameter] = model[parameter].predict(normalized_bufferx)
         means[parameter] = means[parameter].flatten()
@@ -699,9 +763,13 @@ def compute_gp_prediction_mean_and_std(bufferx, model, param_space):
 
         # Precision can sometimes lead GPy to predict extremely low deviation, which leads to numerical issues
         # We add a floor to std to avoid these numerical issues. The majority of std values observed are naturally above this floor.
-        stds[parameter][stds[parameter] < 10**-11] = 10**-11 
+        stds[parameter][stds[parameter] < 10**-11] = 10**-11
+        if var:
+            uncertainty[parameter] = stds[parameter]**2
+        else:
+            uncertainty[parameter] = stds[parameter]
 
-    return means, stds
+    return means, uncertainty
 
 def sample_gp_posterior(bufferx, model, param_space):
     """
